refactor: drop zustand stores in dbInstance and tableData
