+++ frontend/pages/_app.tsx
@@ -1,10 +1,15 @@
+import { DBInstanceContextProvider } from "@/stores";
 // Import TailwindCSS here.
 import "../styles/globals.css";
 import "antd/dist/antd.css";
 import type { AppProps } from "next/app";
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+  return (
+    <DBInstanceContextProvider>
+      <Component {...pageProps} />
+    </DBInstanceContextProvider>
+  );
 }
 
 export default MyApp;
