+++ frontend/types/region.ts
@@ -1,5 +1,4 @@
 import { Term } from "./term";
-import { useDBInstanceStore } from "@/stores/dbInstance";
 
 export type Region = {
   code: string;
@@ -11,18 +10,4 @@ export type AvailableRegion = {
   // providerCode is the mapping between between provider and the region code
   // e.g. N. Virginia is coded 'us-east-1' in AWS,  us-east4 in GCP
   providerCode: Map<string, string>;
-};
-
-export const isValidRegion = (regionList: string[]): boolean => {
-  const availableRegionSet = useDBInstanceStore
-    .getState()
-    .getAvailableRegionSet();
-  for (const idx in regionList) {
-    const region = regionList[idx];
-    if (!availableRegionSet.has(region)) {
-      return false;
-    }
-  }
-
-  return true;
 };
